add textOnlySummary
