+++ packages/server/billing/helpers/sendTeamsLimitEmail.ts
@@ -1,5 +1,5 @@
 import getMailManager from '../../email/getMailManager'
-import limitsEmailCreator from '../../email/limitsEmailCreator'
+import teamLimitsEmailCreator from '../../email/teamLimitsEmailCreator'
 import IUser from '../../postgres/types/IUser'
 
 export type TeamLimitsEmailType = 'locked' | 'sevenDayWarning' | 'thirtyDayWarning'
@@ -15,7 +15,7 @@ type Props = {
 const sendTeamsLimitEmail = (props: Props) => {
   const {user, orgName, orgId, emailType, stickyTeamCount} = props
   const {id: userId, preferredName, email} = user
-  const {subject, body, html} = limitsEmailCreator({
+  const {subject, body, html} = teamLimitsEmailCreator({
     userId,
     orgId,
     preferredName,
